Moved the clean up files to set_up_sql
